Ayush

+++ src/utils/cloudinary.js
@@ -28,11 +28,15 @@ const uploadOnCloudinary = async (localFilePath) => {
         //file has been uploaded successfully
         // console.log("file uploaded on cloudinary", responce.url);
         //
-        await fs.unlinkSync(localFilePath); // unlik file path
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath); // unlik file path
+        }
         return responce; // for user
     } catch (error) {
         console.log("error on cloudinary uploading :", error);
-        await fs.unlinkSync(localFilePath); // will remove local saved file if upload file failed
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath); // will remove local saved file if upload file failed
+        }
         return null;
     }
 };
